Add files via upload
These are the LAMMPS input files (and supporting files), and LAMMPS .data "restart" files for the TGA (thermal ramping) simulations. Not included are the 450 Kelvin versions - but any of these can easily be modified to ramp to 450 Kelvin instead of 150/250/350 etc. - the rest of the code is the same except for the ramping thermostat's final temperature.